backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.1
litellm>=1.43.0
instructor>=1.4.0
neo4j>=5.22.0
graphiti-core>=0.5.0
aiofiles>=23.2.1
tiktoken>=0.7.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0.0
email-validator>=2.0.0
